# Optimize read and write ops