--- a/cli/dispatch_minimal.rs
+++ b/cli/dispatch_minimal.rs
@@ -111,7 +111,7 @@ mod ops {
       None => Box::new(futures::future::err(errors::bad_resource())),
       Some(resource) => Box::new(
         tokio::io::read(resource, zero_copy)
-          .map_err(|err| err.into())
+          .map_err(errors::DenoError::from)
           .and_then(move |(_resource, _buf, nread)| Ok(nread as i32)),
       ),
     }