--- a/cli/dispatch_minimal.rs
+++ b/cli/dispatch_minimal.rs
@@ -3,7 +3,7 @@
 //! Connects to js/dispatch_minimal.ts sendAsyncMinimal This acts as a faster
 //! alternative to flatbuffers using a very simple list of int32s to lay out
 //! messages. The first i32 is used to determine if a message a flatbuffer
-//! message or a "minimal" message, see `has_minimal_token()`.
+//! message or a "minimal" message.
 use crate::state::ThreadSafeState;
 use deno::Buf;
 use deno::Op;
@@ -14,13 +14,9 @@ const DISPATCH_MINIMAL_TOKEN: i32 = 0xCAFE;
 const OP_READ: i32 = 1;
 const OP_WRITE: i32 = 2;
 
-pub fn has_minimal_token(s: &[i32]) -> bool {
-  s[0] == DISPATCH_MINIMAL_TOKEN
-}
-
 #[derive(Copy, Clone, Debug, PartialEq)]
 // This corresponds to RecordMinimal on the TS side.
-struct Record {
+pub struct Record {
   pub promise_id: i32,
   pub op_id: i32,
   pub arg: i32,
@@ -42,26 +38,58 @@ impl Into<Buf> for Record {
   }
 }
 
-impl From<&[i32]> for Record {
-  fn from(s: &[i32]) -> Record {
-    let ptr = s.as_ptr();
-    let ints = unsafe { std::slice::from_raw_parts(ptr, 5) };
-    assert_eq!(ints[0], DISPATCH_MINIMAL_TOKEN);
-    Record {
-      promise_id: ints[1],
-      op_id: ints[2],
-      arg: ints[3],
-      result: ints[4],
-    }
+pub fn parse_min_record(bytes: &[u8]) -> Option<Record> {
+  if bytes.len() % std::mem::size_of::<i32>() != 0 {
+    return None;
   }
+  let p = bytes.as_ptr();
+  #[allow(clippy::cast_ptr_alignment)]
+  let p32 = p as *const i32;
+  let s = unsafe { std::slice::from_raw_parts(p32, bytes.len() / 4) };
+
+  if s.len() < 5 {
+    return None;
+  }
+  let ptr = s.as_ptr();
+  let ints = unsafe { std::slice::from_raw_parts(ptr, 5) };
+  if ints[0] != DISPATCH_MINIMAL_TOKEN {
+    return None;
+  }
+  Some(Record {
+    promise_id: ints[1],
+    op_id: ints[2],
+    arg: ints[3],
+    result: ints[4],
+  })
+}
+
+#[test]
+fn test_parse_min_record() {
+  let buf = vec![
+    0xFE, 0xCA, 0, 0, 1, 0, 0, 0, 2, 0, 0, 0, 3, 0, 0, 0, 4, 0, 0, 0,
+  ];
+  assert_eq!(
+    parse_min_record(&buf),
+    Some(Record {
+      promise_id: 1,
+      op_id: 2,
+      arg: 3,
+      result: 4,
+    })
+  );
+
+  let buf = vec![];
+  assert_eq!(parse_min_record(&buf), None);
+
+  let buf = vec![5];
+  assert_eq!(parse_min_record(&buf), None);
 }
 
 pub fn dispatch_minimal(
   state: &ThreadSafeState,
-  control32: &[i32],
+  record: Record,
   zero_copy: Option<PinnedBuf>,
 ) -> Op {
-  let record = Record::from(control32);
   let is_sync = record.promise_id == 0;
   let min_op = match record.op_id {
     OP_READ => ops::read(record.arg, zero_copy),