--- a/cli/ops.rs
+++ b/cli/ops.rs
@@ -2,6 +2,8 @@
 use atty;
 use crate::ansi;
 use crate::compiler::get_compiler_config;
+use crate::dispatch_minimal::dispatch_minimal;
+use crate::dispatch_minimal::has_minimal_token;
 use crate::errors;
 use crate::errors::{DenoError, DenoResult, ErrorKind};
 use crate::fs as deno_fs;
@@ -74,10 +76,15 @@ fn empty_buf() -> Buf {
   Box::new([])
 }
 
-/// Processes raw messages from JavaScript.
-/// This functions invoked every time Deno.core.dispatch() is called.
-/// control corresponds to the first argument of Deno.core.dispatch().
-/// data corresponds to the second argument of Deno.core.dispatch().
+fn slice_u8_as_i32(bytes: &[u8]) -> &[i32] {
+  let p = bytes.as_ptr();
+  // Assert pointer is 32 bit aligned before casting.
+  assert_eq!((p as usize) % std::mem::align_of::<i32>(), 0);
+  #[allow(clippy::cast_ptr_alignment)]
+  let p32 = p as *const i32;
+  unsafe { std::slice::from_raw_parts(p32, bytes.len() / 4) }
+}
+
 pub fn dispatch_all(
   state: &ThreadSafeState,
   control: &[u8],
@@ -86,6 +93,27 @@ pub fn dispatch_all(
 ) -> Op {
   let bytes_sent_control = control.len();
   let bytes_sent_zero_copy = zero_copy.as_ref().map(|b| b.len()).unwrap_or(0);
+
+  let control32 = slice_u8_as_i32(control);
+  let op = if has_minimal_token(control32) {
+    dispatch_minimal(state, control32, zero_copy)
+  } else {
+    dispatch_all_legacy(state, control, zero_copy, op_selector)
+  };
+  state.metrics_op_dispatched(bytes_sent_control, bytes_sent_zero_copy);
+  op
+}
+
+/// Processes raw messages from JavaScript.
+/// This functions invoked every time Deno.core.dispatch() is called.
+/// control corresponds to the first argument of Deno.core.dispatch().
+/// data corresponds to the second argument of Deno.core.dispatch().
+pub fn dispatch_all_legacy(
+  state: &ThreadSafeState,
+  control: &[u8],
+  zero_copy: Option<PinnedBuf>,
+  op_selector: OpSelector,
+) -> Op {
   let base = msg::get_root_as_base(&control);
   let is_sync = base.sync();
   let inner_type = base.inner_type();
@@ -99,7 +127,6 @@ pub fn dispatch_all(
   let op: Box<OpWithError> = op_func(state, &base, zero_copy);
 
   let state = state.clone();
-  state.metrics_op_dispatched(bytes_sent_control, bytes_sent_zero_copy);
 
   let fut = Box::new(
     op.or_else(move |err: DenoError| -> Result<Buf, ()> {