Essentially side step flatbuffers.

Benching `deno -A tests/http_bench.ts`

this branch:
```
~/src/deno> third_party/wrk/mac/wrk http://127.0.0.1:4500/
Running 10s test @ http://127.0.0.1:4500/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   279.10us  479.38us   8.05ms   96.60%
    Req/Sec    22.82k     1.46k   26.34k    79.21%
  458621 requests in 10.10s, 22.31MB read
Requests/sec:  45403.95
Transfer/sec:      2.21MB
```
master:
```
~/src/deno> third_party/wrk/mac/wrk http://127.0.0.1:4500/
Running 10s test @ http://127.0.0.1:4500/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   380.73us  545.44us   6.88ms   96.20%
    Req/Sec    16.09k     1.70k   18.80k    82.00%
  320396 requests in 10.01s, 15.58MB read
Requests/sec:  32017.73
Transfer/sec:      1.56MB